HDFS-6440. Support more than 2 NameNodes. Contributed by Jesse Yates. (backported by Chao)
Summary:
This is needed for observer namenode. Note this backport only contains part
of the original patch, including:

  1. the changes on tests to support multiple namenodes
  2. changes on BlockTokenSecretManager and BootstrapStandby, to support
     multiple NNs

Notably, this doesn't include changes on EditLogTailer and StandbyCheckpointer
from the original patch. Those are for multiple standby NNs which we don't need.
Also, there are a few bugs in those changes.

Reviewers: ekanth, weiy, ang

Reviewed By: ekanth

Differential Revision: https://code.uberinternal.com/D1456135