DistCp: Change the behavior of "-i" to ignore only the FileNotFoundException
Summary:
With "-i", all read failures are ignored, but normally we only want to ignore the FileNotFoundException.  This diff changes the behavior of the existing option.

I prefer this diff to https://code.uberinternal.com/D688013 because this diff is much easier to manage internally given it's a one-line change.

Reviewers: tejat, haohui, ang, georgeli, O2125 User ekanth: Add reviewers, O2123 User sunchao: Add reviewers, O2129 User mmathew: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, aihuaxu

Reviewed By: O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, aihuaxu

Subscribers: O2131 Project Hadoop-HDFS Dev: Add subscribers, O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers

Differential Revision: https://code.uberinternal.com/D2565961