HDFS observer namenode implementation
Summary:
This adds a new type of namenode: observer. A observer is like a standby
NN (in fact they share most of the code), EXCEPT it doesn't participate
in either NN failover (i.e., it is not part of the HA), or check pointing.

A observer can be specified through configuration. First, it needs to be
added into the config: dfs.ha.namenodes, just like a normal namenode,
together with other configs such as dfs.namenode.rpc-address,
dfs.namenode.http-address, etc. Second, it needs to be specified in a new
config: dfs.ha.observer.namenodes. This differentiate it from the ordinary
active/standby namenodes.

A observer can be used to serve stale read-only requests from HDFS client,
when the following two conditions are satisfied:

  1. the config dfs.client.failover.proxy.provider.<nameservice> is
     set to org.apache.hadoop.hdfs.server.namenode.ha.StaleReadProxyProvider.
  2. the config dfs.client.enable.stale-read is set to true

This also changes the way edit logs are loaded from the standby/observer NNs.
Instead of loading them all at once, the new implementation loads them
one batch at a time (default batch size is 10K edits) through multiple
iterations, while waiting for a short amount of time in between the
iterations (default waiting time is 100ms). This is to make sure the global
lock won't be held too long during loading edits. Otherwise, the RPC
processing time would suffer.

This also added a separate configuration for specifying the edit log interval
for observers, since it may need to perform this action in a much higher
frequency. The related configuration is dfs.ha.observer.tail-edits.ms.period.

This patch does not include a mechanism for clients to specify the bound of
the staleness using journal transction ID: excluding this allows us to
deploy the observer more easily. Instead, it provides a configuration:
dfs.ha.observer.stale.limit.ms, which specifies the maximum lagging between
observer and active. If the actual lagging reaches this threshold, requests
will be forwarded to the other observers and active if none is available.

Potential future tasks:

  1. use journal transaction ID to ensure bound on staleness. This can be
     embedded in the RPC header.
  2. allow new standby/observer to be deployed without datanode restart.

Reviewers: weiy, kelvin, ang, ekanth

Reviewed By: ekanth

Subscribers: link

Differential Revision: https://code.uberinternal.com/D1458279