[ObserverNameNode] retry active when FNF in open, getListing & getFileInfo
Summary:
This modifies the ObserverReadProxyProvider to retry active namenode
when encountered FNF exception in open(), getListing() and getFileInfo() calls.

The way how FNF exception is thrown for the above calls are different: for open(),
it is thrown from the server side and is wrapped in a RemoteException; for others,
the server side simply returns null but the FNF exception is thrown on the client
side after checking the result. This commit checks both cases.

Reviewers: ekanth, weiy, ang

Reviewed By: ekanth, weiy

Differential Revision: https://code.uberinternal.com/D1600519