Consider observer reads flag when caching FileSystem
Summary:
FileSystem in default is cached for reuse. However this caused a
problem for reading from Observer: suppose one change the config
and enable/disable the flag, the change will not take effect since
the cached FileSystem is used.

In this commit, we added an extra field in the cache key class, for
observer read. In this way, the caching would consider that and
return different FileSystem objects accordingly.

Reviewers: ekanth, ang, crh

Reviewed By: ekanth

Maniphest Tasks: T1748107

Differential Revision: https://code.uberinternal.com/D1743579