HDFS-13602. Add checkOperation(WRITE) checks in FSNamesystem. Contributed by Chao Sun.
Summary:
(cherry picked from commit ff013d2c952272f3176dcf624251b05d610503b5)

HDFS-13602:
Similar to the work done in HDFS-4591 to avoid having to take a write lock before checking if an operation category is allowed, we can do the same for the write lock that is taken sometimes (when updating access time) within getBlockLocations.

This is particularly useful when using the standby read feature (HDFS-12943), as it will be the case on an observer node that the operationCategory(READ) check succeeds but the operationCategory(WRITE) check fails. It would be ideal to fail this check before acquiring the write lock.

Test Plan: ETA

Reviewers: aihuaxu, O2123 User sunchao: Add reviewers, O2125 User ekanth: Add reviewers, O2127 User pgolash: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers

Reviewed By: aihuaxu, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers

Subscribers: O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers, O2131 Project Hadoop-HDFS Dev: Add subscribers

Differential Revision: https://code.uberinternal.com/D5093293