[ObserverNameNode] throw retriable exception when observer is in safemode
Summary:
When client issues getBlockLocations() requests to a observer that
is in safemode, it will sometimes get SafeModeException if no blocks for
the target file have been received on the observer side. This will cause
client requests to fail, which is undesirable. Instead, it should return
RetriableException so clients can retry other observer/active namenodes.
This logic is used for active namenode but not observer right now.

Reviewers: ekanth, crh, jeffreyz

Reviewed By: ekanth

Maniphest Tasks: T1778127

Differential Revision: https://code.uberinternal.com/D2041811