[ObserverNameNode] Retry active when encounter BlockMissingException
Summary:
Sometimes RPCs to observer namenode may fail with "BlockMissingException".
This may due to that we have a split of DNs between active and observer,
or that block reports came very late to observer.

The DFSInputStream already has logic to retry "namenode.getBlockLocations"
upon such failures. In observer case, we should retry active when this
happens.

Test Plan: unit tests

Reviewers: ekanth, crh, balaga, jeffreyz

Reviewed By: ekanth

Differential Revision: https://code.uberinternal.com/D2057227