DistCp: Widen the rules for -i.
Summary: We found out that -i doesn't work for the case that the file gets deleted between listing status and actual copying the file inside the CopyMapper.map() function, because the exception had 2 levels of IOException.  This diff will fix that.

Test Plan: Compile and test with -listing and -norefresh

Reviewers: O2125 User ekanth: Add reviewers, O2123 User sunchao: Add reviewers, O2129 User mmathew: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, aihuaxu

Reviewed By: O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, aihuaxu

Subscribers: O2131 Project Hadoop-HDFS Dev: Add subscribers, O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers, jenkins

Differential Revision: https://code.uberinternal.com/D2566629