Fix router client retry in observer router
Summary:
In current RouterRpcClient retry logic, it will directly retry when no active namenode
is available for the nameservice for the first time and then directly fail for the following
retries without actual retry.
This doesn't hold true anymore since in Observer Router, it is possible that a name service
has no active namenode and all are observer namenodes. Thus we need to handle this new case.

Test Plan: Unit test to make sure the logic is complete for observer router

Reviewers: O2125 User ekanth: Add reviewers, O2123 User sunchao: Add reviewers, O2127 User pgolash: Add reviewers, O2129 User mmathew: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, crh

Reviewed By: O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, crh

Subscribers: O2131 Project Hadoop-HDFS Dev: Add subscribers, O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers

Maniphest Tasks: T4107665

Differential Revision: https://code.uberinternal.com/D3383559