Continue making hadoop scripts taking more env scripts
Summary:
This diff is a follow up on https://code.uberinternal.com/D3443199, which
introduces the env var to the hadoop-config.sh.
This diff focus on the callers of the hadoop-config, including hdfs and
hadoop.
This diff by no means to fix all of the issues in these scripts. i.e
hdfs has some unused options (--config).

Test Plan:
Tested on observers, both
./scripts/nn/start.sh + ./scripts/observer/start.sh
and
./scripts/zing-nn/start.sh + ./scripts/zing-observer/start.sh
can work

Reviewers: O2123 User sunchao: Add reviewers, O2125 User ekanth: Add reviewers, O2127 User pgolash: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, liangg

Reviewed By: O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, liangg

Subscribers: O2131 Project Hadoop-HDFS Dev: Add subscribers, O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers

Differential Revision: https://code.uberinternal.com/D3449987