Expose aggregated memory and vcores used by an application
Summary: allocatedMemory and allocatedVcores will be -1 after the application is done and actually it should be used to track the running memory and vcores. The patch exposes aggregated used memory and used vcores to  aggregatedAllocatedMemory and aggregatedAllocatedVcores so we can track how much memory and vcores are used in a certain time period.

Test Plan: test locally and unit tests.

Reviewers: O2123 User sunchao: Add reviewers, O2125 User ekanth: Add reviewers, O2127 User pgolash: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, pgolash, tracyliu

Reviewed By: O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, tracyliu

Subscribers: O2131 Project Hadoop-HDFS Dev: Add subscribers, O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers

Maniphest Tasks: T4699881, T4670885

Differential Revision: https://code.uberinternal.com/D3697601