HDFS-15118. [SBN Read] Slow clients when Observer reads are enabled but there are no Observers on the cluster. Contributed by Chen Liang
Summary:
From HDFS-15118:

Substantial degradation in performance of HDFS clients, when Observer reads are enabled via ObserverReadProxyProvider, but there are no ObserverNodes on the cluster.

We have 4 NameNodes on a very busy cluster - one Active and three Standbys. In this setup the degradation of performance is particularly seen on the localization stage of jobs that have a lot of jar dependencies.
The main reasons for the slowness are

ObserverReadProxyProvider tries to find ObserverNode for every read call. See ObserverReadInvocationHandler.invoke(). invoke() skips all proxies that are not Observer, including Active. It calls getHAServiceState() for each of the NameNodes. If it doesn't find an Observer, it will use failoverProxy to send the call, which points to the Active. So any read call when there is no Observers translates into 5 RPCs making clients slow.
StandbyNode treats getHAServiceState() as a READ operation and throws StandbyException. I think processing an exception instead of a normal response contributes to the slowness. As 3 out of 5 calls carry the StandbyException.
Fixing the problem should include two changes.

Make ORPP sticky. If it chose Active as currentProxy, it should keep sending reads to it for some configurable period of time (10 min). Then it can try to switch to an Observer.
Allow getHAServiceState() on StandbyNode without throwing StandbyException.

Test Plan: ETA

Reviewers: aihuaxu, O2123 User sunchao: Add reviewers, O2125 User ekanth: Add reviewers, O2127 User pgolash: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers

Reviewed By: aihuaxu, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers

Subscribers: O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers, O2131 Project Hadoop-HDFS Dev: Add subscribers

Differential Revision: https://code.uberinternal.com/D5074245