POC patch for treating abusive user with fair call queue
Summary:
This first patch introduced a blacklist concept which is controlled
by a feature flag.
Normally, fair call queue will put heavy users in the queue with lower
priority whereas with the blacklist, only users in the blacklist will
be dropped in these queues. This is the first step to onboard fair call
queue into our current namenode with the expecation of minimum change
for traffic handling.
Initially users will be manually identified and added to this blacklist.
Later we can turn off the blacklist feature flag and start to really use
fair call queue (with white list feature in a following patch)

Test Plan: UT + staging test (to be done)

Reviewers: O2123 User sunchao: Add reviewers, O2125 User ekanth: Add reviewers, O2127 User pgolash: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, sunchao

Reviewed By: O2123 User sunchao: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, sunchao

Subscribers: sunchao, liangg, O2131 Project Hadoop-HDFS Dev: Add subscribers, O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers

Differential Revision: https://code.uberinternal.com/D3906787