YARN RM Webservices changes to filter scheduler info by queue and exposes partition resource.
Summary:
We are exposing queue level scheduler info in this patch and additionally, we add absolute
partition resource for the partition. This can be used to query partition utilization and partition
capacity.

Currently, YARN only returns full scheduler info for an endpoint like this:
http://hadoopbenchaqjm01-sjc1.prod.uber.internal:8088/ws/v1/cluster/scheduler

This returns full data, which is sometimes too much.

Added support for -

http://hadoopbenchaqjm01-sjc1.prod.uber.internal:8088/ws/v1/cluster/scheduler/queue=test123

Test Plan: unit test and test locally

Reviewers: O2123 User sunchao: Add reviewers, O2125 User ekanth: Add reviewers, O2127 User pgolash: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, pgolash, tracyliu

Reviewed By: O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, tracyliu

Subscribers: O2131 Project Hadoop-HDFS Dev: Add subscribers, O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers

Differential Revision: https://code.uberinternal.com/D3999029