Add Top Metrics for Files in Listing ops
Summary:
This adds "# of files in listStatus calls" to `TopMetrics`. Currently
we only have metrics for total number of files across all listStatus calls.
With this, it is hard to identify which individual user contribute most
to the listing overhead. Therefore, it will be useful to have the TopN users
available for debugging.

Reviewers: liangg, lfengnan, O2125 User ekanth: Add reviewers, O2127 User pgolash: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers

Reviewed By: liangg, lfengnan, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers

Subscribers: O2131 Project Hadoop-HDFS Dev: Add subscribers, O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers

Differential Revision: https://code.uberinternal.com/D4353335