Node transition changes to avoid killing of containers when node is unhealthy due to disk usage
Summary:
Re-used the stressed node framework to plug one more state change i.e
a) If the node is unhealthy due to all disks > 95%, avoid killing of containers and don't let any new job schedule on it

This is required to mitigate a lot of MA job failures to some extent (This will give some opportunity to let containers finish and free the disk).

As of now, NM is **not** emitting disk usage per containers, else we could extend the framework to kill containers using the most disk.  (I will create a ticket to see if this is do-able as part of the feedback to RM)

Another approach is not to make RM changes, but let the threshold of disk = 100% on NM. This pose some risk since RM change is config controlled and it will revert to old behavior of killing once more NMs > 95%.

Test Plan: Unit test cases

Reviewers: mabansal, tracyliu, zshuyi, #yarn_efforts, O2123 User sunchao: Add reviewers, O2125 User ekanth: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers

Reviewed By: tracyliu, zshuyi, #yarn_efforts, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers

Subscribers: O2131 Project Hadoop-HDFS Dev: Add subscribers, O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers

Differential Revision: https://code.uberinternal.com/D4421949