HADOOP-17068. Client fails forever when namenode ipaddr changed. Contributed by Sean Chow.
Summary:
For machine replacement, if you replace the standby namenode with a new ipaddr but keep the same hostname and then failover to the new namenode, the client will fail to read or write forever until it's restarted. That is because the code handleConnectionFailure() throws an exception and failovers tothe next namenode. That doesn't give the chance to retry if the ipAddress of the namenode changes.

This patch ignores the exception if the ipAddress changes when the exception occurs to let the client retry.

Test Plan: Test on staging

Reviewers: O2125 User ekanth: Add reviewers, O2127 User pgolash: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, liangg, ekanth

Reviewed By: O2125 User ekanth: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, liangg, ekanth

Subscribers: O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers, O2131 Project Hadoop-HDFS Dev: Add subscribers

JIRA Issues: HDFS-285

Differential Revision: https://code.uberinternal.com/D4947153