Support detecting NameNode IP change smoothly in DataNode
Summary:
NameNode is critical in HDFS cluster and in some cases we have to replace NameNode
due to hardware or performance issues. Typically we take hostname rename (NameNode IP address change)
approach to mimimize the client impact, while DataNodes in HDFS have issues to detect such changes:
the DataNode does retry with the new IP but it doesn't remember the new IP for the next connection,
so for each connection to NameNode, e.g., heartbeats, it succeeds in the second retry.

The change is to update new IP address within remoteId that is shared across all the connections, in
addition to the local copy ('server') within Client. With that, only the first retry fails and the following
will succeed without retry.

Test Plan: Tested with local setup. Unfortunately there is no easier way to add unit test.

Reviewers: O2125 User ekanth: Add reviewers, O2127 User pgolash: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, lfengnan, liangg, zhangf, mmathew, mmatthews

Reviewed By: O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, liangg, mmathew

Subscribers: jeffreyz, mmathew, O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers, O2131 Project Hadoop-HDFS Dev: Add subscribers

JIRA Issues: HDFS-631

Differential Revision: https://code.uberinternal.com/D5336099