[HDP] Allow passing in the NEW_VERSION using the jenkins RELEASE_VERSION parameter
Summary: Without this, the script is calculating the version number as 164. But unfortunately, 164 is already pushed and thus cannot be replaced. We furthermore don't support SNAPSHOT jars. This is an attempt to work around this by publishing a user specified version.

Test Plan: Successfully built a test jar using https://ci.uberinternal.com/job/hadoop-common-release/113/console with my name on it.

Reviewers: jingzhao, lfengnan, mmathew, aihuaxu, O2123 User sunchao: Add reviewers, O2125 User ekanth: Add reviewers, O2127 User pgolash: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers

Reviewed By: lfengnan, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers

Subscribers: jeffreyz, O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers, O2131 Project Hadoop-HDFS Dev: Add subscribers

JIRA Issues: HDP-18259

Differential Revision: https://code.uberinternal.com/D5858423