[HDP] Change version string to conform to existing code expectations
Summary:
It turns out that a lot of existing code like [1] and [2] expect the hadoop version info to be Num.Num.Num style, and they break with a prefix like uber-hadoop :-). There is no strong reason to persist with the uber-hadoop prefix, and infact the prefix uber-hadoop is arguably redundant, if the git hash is burned in.

[1] https://sourcegraph.uberinternal.com/code.uber.internal/data/hive/-/blob/shims/common/src/main/java/org/apache/hadoop/hive/shims/ShimLoader.java?subtree=true#L143
[2] https://sourcegraph.uberinternal.com/code.uber.internal/data/orc/-/blob/java/core/src/java/org/apache/orc/impl/HadoopShimsFactory.java#L50

Changing this version string is required since otherwise it will break at least Hive and Spark but also maybe Presto.

Test Plan: Live

Reviewers: mmathew, jingzhao, lfengnan, O2123 User sunchao: Add reviewers, O2125 User ekanth: Add reviewers, O2127 User pgolash: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers

Reviewed By: mmathew, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers

Subscribers: O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers, jeffreyz, O2131 Project Hadoop-HDFS Dev: Add subscribers

JIRA Issues: HDP-18259

Differential Revision: https://code.uberinternal.com/D5864321