Add code path for fast switch
Summary:
This will be the patch that allows us to start testing.

We now use an ExecutorCompletionService to execute doRead() and give a timeout value for the poll(). When timeout reaches, we decide to switch to another reader and interrupt the current one to let it die.

To avoid too many retries, we give a maxSwitch value for each inputStream.
To avoid client-side slowness causing complete read failure, when we find all datanodes are "slow" we give up any further switch.

TODO :
- Add more testings, especially for regression.
- Organize code and fix styles.

Reviewers: O2123 User sunchao: Add reviewers, O2125 User ekanth: Add reviewers, O2127 User pgolash: Add reviewers, O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, jingzhao, zhangf, aihuaxu

Reviewed By: O2121 Project Hadoop-HDFS Dev: Add blocking reviewers, jingzhao

Subscribers: mmathew, O2133 Project Data Foundation - Hadoop Ecosystem: Add subscribers, O2131 Project Hadoop-HDFS Dev: Add subscribers

JIRA Issues: HDFS-1138

Differential Revision: https://code.uberinternal.com/D5846741